+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.4 (Unreleased)
 
 - fix: KCC characters getting stuck when surrounded by steep slopes due to penetration recovery velocity being included in vertical wall constraints
+- fix: incorrect closest-point-on-edge in convex hull builder, increase coplanar tolerance to fix degenerate hulls
 - feat: defer DOF masking in contact solver to single write-back per constraint
 - feat: reduce convex radius if too big for shape
 - feat: added `debugRenderer.dispose` method to `crashcat/three` debug renderer utility